Update test_fragile_finder.py
I still didn't really understand the purpose of the first few functions, or how to use pytest correctly as the example described uses commandline inputs and I couldn't see how that was required for my scripts, so I wrote my own functionality tests for the two main functions, parse_blast and remove_duplicates, which have a test file and an expected file and ensure that the test file can produce the expected output. Is this what was meant by functionality tests?